strong checks

+++ include/C/fileutil.h
@@ -12,4 +12,4 @@ fileDesc getFileDesc(FILE *file);
 
 void freeFileDesc(fileDesc *fileD);
 
-size_t getFileSize(FILE *file);
+long getFileSize(FILE *file);
